Update specie single page and fish model in angular. Update controllers
From species is now possible to remove fish from the list. This functionality will be removed because fishes must have one specie, but it will be useful for other entities as reference. Update angular model references: instead of giving them an object, they will receive a string or number. This is necessary because, after fixing infinite recursion on bidirectional relationship, the JSON REST responses start including numbers or string instead of full objects.