Included single fish page to update data
Is now possible to open a single page containing the information of a single fish and make changes. FishController update method has been changed too. Now it receives a HashMap of values instead of a Fish object. That is necessary because the specie/pool received are not object but only the name (for specie) and id (for pool), so it's not possible to deserialize data in a Fish object.