Add JWT decoder. Update login and logout
JWT decoder is used to extract informations from the token. Now every token carry the staff id, staff role, expire time, the creation time and the username. During the registration, it's required to fill other fields for generating an object Staff, which will be assigned in a unilateral relation to the object User.

Login will store the token in local storage. Logout remove it from local storage. By now, it's not checked if the token has expired.

WIP for admin section which is still not working for both logged users or not logged users. Need to fix all the services.