Fixed security configurations, models and controllers.
The unidirectional relation between User and Staff is now a bidirectional relation. It helps to manage the authentication after deleting a staff member.

StaffController now allows to lists the credentials. This path is locked. 

Authentication by token will fail if the Staff entity referenced by the token is not anymore in db. 

User password is hashed. Implemented a method to verify the correcteness of login 
 passwords.