Created error notifications.
Every error in admin page will be notified with a message. Clicking it will remove the error. If the error is of type 401, than the user will be redirect to homepage and forced to logout. This will grants that, after removing a staff, the related user will be also disabilitated to access the admin page.