Disable removal of admins. Update registration of staff.
Admins can now be removed only if in a number greater than 1 in the database.

Staff can be removed if there are no responsabilities between sectors, pools and activities.

Registration is available only for the first time, after that only a staff can registrate others staff members. 

Update authenticationService registration method in order to create a correct staff registration depending on who is requesting.

Update Login HTML. Is now possible to swap between registration and login.